Add missing function call in readme
The onUpdate() function in the skills file was never called (as far as I could tell), so the update was never sent up to the body's ajax call. I've added a call to onUpdate inside the handleEdit function in the skills file. I'm guessing this is where people were getting stuck.